Moved state into separate module
