+++ func/demo/CMakeLists.txt
@@ -11,7 +11,7 @@ faasm_func(net net.cpp)
 faasm_func(noop noop.cpp)
 faasm_func(print print.cpp)
 faasm_func(sizes sizes.cpp)
-faasm_func(state state.cpp)
+faasm_func(state_sync state_sync.cpp)
 faasm_func(state_async state_async.cpp)
 faasm_func(state_offset state_offset.cpp)
 faasm_func(stateop stateop.cpp)
