+++ include/wasm/wasm.h
@@ -29,162 +29,6 @@ namespace wasm {
 
     Uptr getNumberOfPagesForBytes(U32 nBytes);
 
-    typedef std::pair<long, long> Segment;
-    typedef std::set<std::pair<long, long>> SegmentSet;
-
-    /**
-     * Wrapper around segment of shared memory
-     */
-    class StateMemorySegment {
-    public:
-        StateMemorySegment(Uptr offsetIn, uint8_t *ptrIn, size_t lengthIn);
-
-        Uptr offset;
-        uint8_t *ptr;
-        size_t length;
-        bool inUse;
-    };
-
-    /**
-     * A wrapper around shared memory for state
-     */
-    class StateMemory {
-    public:
-        explicit StateMemory(const std::string &user);
-
-        ~StateMemory();
-
-        uint8_t *createSegment(size_t length);
-
-        void releaseSegment(uint8_t *ptr);
-
-        Runtime::GCPointer<Runtime::Memory> wavmMemory;
-        Runtime::GCPointer<Runtime::Compartment> compartment;
-    private:
-        std::shared_mutex memMutex;
-
-        Uptr nextByte;
-        const std::string user;
-        std::vector<StateMemorySegment> segments;
-    };
-
-    /**
-     * A specific key/value pair for state
-     */
-    class StateKeyValue {
-    public:
-        explicit StateKeyValue(const std::string &keyIn, size_t sizeIn);
-
-        StateKeyValue(const StateKeyValue& other);
-
-        StateKeyValue& operator=(const StateKeyValue& other);
-
-        const std::string key;
-
-        void get(uint8_t *buffer);
-
-        void getSegment(long offset, uint8_t *buffer, size_t length);
-
-        void set(uint8_t *buffer);
-
-        void setSegment(long offset, uint8_t *buffer, size_t length);
-
-        void pull(bool async);
-
-        void pushFull();
-
-        void pushPartial();
-
-        void clear();
-
-        void lockRead();
-
-        void unlockRead();
-
-        void lockWrite();
-
-        void unlockWrite();
-
-        bool empty();
-
-    private:
-        std::atomic<bool> isWholeValueDirty;
-
-        std::vector<bool> dirtyFlags;
-        std::shared_mutex dirtySegmentsMutex;
-
-        std::shared_mutex valueMutex;
-
-        util::TimePoint lastPull;
-        util::TimePoint lastInteraction;
-        long staleThreshold;
-        long idleThreshold;
-
-        std::vector<uint8_t> value;
-        size_t size;
-
-        std::atomic<bool> _empty;
-
-        void doRemoteRead();
-
-        void updateLastInteraction();
-
-        bool isStale(const util::TimePoint &now);
-
-        bool isIdle(const util::TimePoint &now);
-    };
-
-    typedef std::map<std::string, StateKeyValue *> KVMap;
-    typedef std::pair<std::string, StateKeyValue *> KVPair;
-
-    /**
-     * Holds state for a given user
-     */
-    class UserState {
-    public:
-        explicit UserState(const std::string &userIn);
-
-        ~UserState();
-
-        StateKeyValue *getValue(const std::string &key, size_t size);
-
-        void pushAll();
-    private:
-        const std::string user;
-
-        KVMap kvMap;
-        std::shared_mutex kvMapMutex;
-    };
-
-    typedef std::map<std::string, UserState *> UserStateMap;
-    typedef std::pair<std::string, UserState *> UserStatePair;
-
-    /**
-     * Global entry point into state
-     */
-    class State {
-    public:
-        State();
-
-        ~State();
-
-        StateKeyValue *getKV(const std::string &user, const std::string &key, size_t size);
-
-        UserState *getUserState(const std::string &user);
-
-        void pushLoop();
-
-        void pushAll();
-
-    private:
-        UserStateMap userStateMap;
-        std::shared_mutex userStateMapMutex;
-
-        long pushInterval;
-    };
-
-    State &getGlobalState();
-
     struct RootResolver : Runtime::Resolver {
         explicit RootResolver(Runtime::Compartment *compartment) {
             Intrinsics::Module &moduleRef = getIntrinsicModule_env();
