+++ lib/CMakeLists.txt
@@ -11,7 +11,10 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fpic -fvisibility=default")
 # For some incredibly annoying reason, CMake isn't picking this up from the environment
 # It's crucial for getting things to link properly so unfortunately must be set here
 if (CMAKE_BUILD_TYPE MATCHES "wasm")
+    add_definitions(-DWASM_BUILD=1)
     set(CMAKE_RANLIB /toolchain/bin/llvm-ranlib)
+else()
+    add_definitions(-DWASM_BUILD=0)
 endif()
 
 add_library(faasm STATIC
