+++ lib/libfaasm.imports
@@ -1,7 +1,6 @@
 __faasm_push_state
 __faasm_push_state_partial
 __faasm_read_state
-__faasm_read_state_ptr
 __faasm_write_state
 __faasm_write_state_offset
 __faasm_read_state_offset
