+++ src/emulator/emulator.cpp
@@ -1,42 +1,65 @@
 #include <faasm/memory.h>
 #include <util/util.h>
 #include <infra/infra.h>
+#include <state/state.h>
 
 /**
- * Dummy implementations of state calls to keep compiler happy
+ * Local implementations of faasm calls
  */
 
-extern "C" {
+state::StateKeyValue *getKv(const char *key, size_t size) {
+    state::State &s = state::getGlobalState();
+    state::StateKeyValue *kv = s.getKV("demo", key, size);
+    return kv;
+}
 
 void __faasm_read_state(const char *key, unsigned char *buffer, long bufferLen, int async) {
-    infra::Redis *redis = infra::Redis::getThreadState();
-    redis->get(key, buffer, bufferLen);
+    state::StateKeyValue *kv = getKv(key, bufferLen);
+
+    bool isAsync = async == 1;
+    kv->pull(isAsync);
+
+    kv->get(buffer);
 }
 
-uint8_t *__faasm_read_state_ptr(const char *key, long len, int async) {
-    throw std::runtime_error("Not implemented");
+void __faasm_read_state_offset(const char *key, long totalLen, long offset, unsigned char *buffer, long bufferLen,
+                               int async) {
+    state::StateKeyValue *kv = getKv(key, totalLen);
+
+    bool isAsync = async == 1;
+    kv->pull(isAsync);
+
+    kv->getSegment(offset, buffer, bufferLen);
 }
 
-void __faasm_write_state(const char *key, const unsigned char *data, long dataLen, int async) {
-    infra::Redis *redis = infra::Redis::getThreadState();
+void __faasm_write_state(const char *key, const uint8_t *data, long dataLen, int async) {
+    state::StateKeyValue *kv = getKv(key, dataLen);
+    kv->set(data);
 
-    if (dataLen > 0) {
-        redis->set(key, data, (size_t) dataLen);
+    if (async == 0) {
+        kv->pushFull();
     }
 }
 
 void __faasm_write_state_offset(const char *key, long totalLen, long offset, const unsigned char *data, long dataLen,
                                 int async) {
-    infra::Redis *redis = infra::Redis::getThreadState();
-    redis->setRange(key, offset, data, dataLen);
+    state::StateKeyValue *kv = getKv(key, totalLen);
+
+    kv->setSegment(offset, data, dataLen);
+
+    if (async == 0) {
+        kv->pushPartial();
+    }
 }
 
-void __faasm_read_state_offset(const char *key, long totalLen, long offset, unsigned char *buffer, long bufferLen,
-                               int async) {
-    infra::Redis *redis = infra::Redis::getThreadState();
+void __faasm_push_state(const char *key) {
+    state::StateKeyValue *kv = getKv(key, 0);
+    kv->pushFull();
+}
 
-    // Importantly getRange is *inclusive*, so we need to knock one off the end of the range
-    redis->getRange(key, buffer, bufferLen, offset, offset + bufferLen - 1);
+void __faasm_push_state_partial(const char *key) {
+    state::StateKeyValue *kv = getKv(key, 0);
+    kv->pushPartial();
 }
 
 long __faasm_read_input(unsigned char *buffer, long bufferLen) {
@@ -44,14 +67,6 @@ long __faasm_read_input(unsigned char *buffer, long bufferLen) {
     std::copy(inputData.begin(), inputData.end(), buffer);
 
     return 14;
-}
-
-void __faasm_push_state(const char *key) {
-
-}
-
-void __faasm_push_state_partial(const char *key) {
-
 }
 
 void __faasm_lock_state_read(const char *key) {
@@ -78,5 +93,4 @@ void __faasm_chain_function(const char *name, const unsigned char *inputData, lo
 
 }
 
-}
