+++ src/state/state.cpp
@@ -1,91 +1,7 @@
-#include <prof/prof.h>
 #include <algorithm>
-#include "wasm.h"
+#include "state.h"
 
-namespace wasm {
-    typedef std::unique_lock<std::shared_mutex> FullLock;
-    typedef std::shared_lock<std::shared_mutex> SharedLock;
-
-    /**
-     * Shared memory segment
-     */
-    StateMemorySegment::StateMemorySegment(Uptr offsetIn, uint8_t *ptrIn, size_t lengthIn) : offset(offsetIn),
-                                                                                             ptr(ptrIn),
-                                                                                             length(lengthIn) {
-
-        inUse = true;
-    }
-
-    /**
-     * Shared memory
-     */
-    StateMemory::StateMemory(const std::string &userIn) : user(userIn) {
-        compartment = Runtime::createCompartment();
-
-        // Prepare memory, each user can have up to 1000
-        IR::MemoryType memoryType(true, {0, 1000});
-
-        // Create a shared memory for this user
-        wavmMemory = Runtime::createMemory(compartment, memoryType, user + "_shared");
-
-        // Create some space initially
-        Runtime::growMemory(wavmMemory, 5);
-
-        nextByte = 0;
-    }
-
-    StateMemory::~StateMemory() {
-        wavmMemory = nullptr;
-
-        Runtime::tryCollectCompartment(std::move(compartment));
-    };
-
-    uint8_t *StateMemory::createSegment(size_t length) {
-        const std::shared_ptr<spdlog::logger> logger = util::getLogger();
-
-        // Need to lock the whole memory to make changes
-        FullLock lock(memMutex);
-
-        Uptr bytesToAdd = (Uptr) length;
-        Uptr thisStart = nextByte;
-        nextByte += bytesToAdd;
-
-        // See if we need to grow
-        Uptr requiredPages = getNumberOfPagesForBytes(nextByte);
-        const Uptr currentPageCount = getMemoryNumPages(wavmMemory);
-        Uptr maxPages = getMemoryMaxPages(wavmMemory);
-        if (requiredPages > maxPages) {
-            logger->error("Allocating {} pages of shared memory when max is {}", requiredPages, maxPages);
-            throw std::runtime_error("Attempting to allocate more than max pages of shared memory.");
-        }
-
-        // Grow memory if required
-        if (requiredPages > currentPageCount) {
-            Uptr expansion = requiredPages - currentPageCount;
-            growMemory(wavmMemory, expansion);
-        }
-
-        // Return pointer to memory
-        U8 *ptr = Runtime::memoryArrayPtr<U8>(wavmMemory, thisStart, length);
-
-        // Record the use of this segment
-        segments.emplace_back(StateMemorySegment(thisStart, ptr, length));
-
-        return ptr;
-    }
-
-    void StateMemory::releaseSegment(uint8_t *ptr) {
-        // Lock the memory to make changes
-        FullLock lock(memMutex);
-
-        // TODO make this more efficient
-        for (auto s : segments) {
-            if (s.ptr == ptr) {
-                s.inUse = false;
-                break;
-            }
-        }
-    }
+namespace state {
 
     /**
      * Key/value
@@ -217,7 +133,7 @@ namespace wasm {
         std::copy(value.data() + offset, value.data() + offset + length, buffer);
     }
 
-    void StateKeyValue::set(uint8_t *buffer) {
+    void StateKeyValue::set(const uint8_t *buffer) {
         this->updateLastInteraction();
 
         // Unique lock for setting the whole value
@@ -235,7 +151,7 @@ namespace wasm {
         _empty = false;
     }
 
-    void StateKeyValue::setSegment(long offset, uint8_t *buffer, size_t length) {
+    void StateKeyValue::setSegment(long offset, const uint8_t *buffer, size_t length) {
         this->updateLastInteraction();
 
         // Check we're in bounds
@@ -403,6 +319,8 @@ namespace wasm {
         for (const auto &iter: kvMap) {
             delete iter.second;
         }
+
+        kvMap.clear();
     }
 
     StateKeyValue *UserState::getValue(const std::string &key, size_t size) {
@@ -462,6 +380,8 @@ namespace wasm {
         for (const auto &iter: userStateMap) {
             delete iter.second;
         }
+
+        userStateMap.clear();
     }
 
     void State::pushLoop() {
@@ -478,6 +398,14 @@ namespace wasm {
         for (const auto &iter: userStateMap) {
             iter.second->pushAll();
         }
+    }
+
+    void State::forceClearAll() {
+        for (const auto &iter: userStateMap) {
+            delete iter.second;
+        }
+
+        userStateMap.clear();
     }
 
     StateKeyValue *State::getKV(const std::string &user, const std::string &key, size_t size) {
