+++ test/lib/test_matrix.cpp
@@ -4,6 +4,7 @@
 #include "faasm/matrix.h"
 
 #include <infra/infra.h>
+#include <state/state.h>
 
 #include <iostream>
 
@@ -30,6 +31,7 @@ namespace tests {
 
     TEST_CASE("Test matrix to redis round trip", "[matrix]") {
         redisQueue.flushAll();
+        state::getGlobalState().forceClearAll();
 
         MatrixXd mat = buildDummyMatrix();
 
@@ -50,6 +52,7 @@ namespace tests {
 
     TEST_CASE("Test updating matrix element in state", "[matrix]") {
         redisQueue.flushAll();
+        state::getGlobalState().forceClearAll();
 
         MatrixXd mat = buildDummyMatrix();
 
@@ -83,6 +86,7 @@ namespace tests {
 
     TEST_CASE("Test reading columns from state", "[matrix]") {
         redisQueue.flushAll();
+        state::getGlobalState().forceClearAll();
 
         long nRows = 4;
         long nCols = 5;
@@ -215,6 +219,7 @@ namespace tests {
 
     TEST_CASE("Test sparse matrix round trip", "[matrix]") {
         redisQueue.flushAll();
+        state::getGlobalState().forceClearAll();
 
         SparseMatrix<double> mat = faasm::randomSparseMatrix(5, 10, 0.4);
 
@@ -229,6 +234,7 @@ namespace tests {
 
     void checkSparseMatrixRoundTrip(int rows, int cols, int colStart, int colEnd) {
         redisQueue.flushAll();
+        state::getGlobalState().forceClearAll();
 
         const char *key = "sparse_trip_offset_test";
